Split Initializer into 2 source files to allow 32-bit debug build
